@@ -1,11 +1,10 @@
-const express = require('express')
+import { Router } from 'express'
+import eventRouter from './event.router'
+import visitorRouter from './visitor.router'
 
-const router = express.Router()
+const router = Router()
 
-router.use(function(req, res, next) {
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-  next();
-}); 
+router.use(eventRouter)
+router.use(visitorRouter)
 
 module.exports = router
