@@ -120,7 +120,7 @@ export async function setContext(app, context) {
       payload: context.payload,
       error: context.error,
       base: '/',
-      env: {}
+      env: {"basePath":"","baseURL":"http://localhost:3005/"}
     }
     // Only set once
     if (context.req) app.context.req = context.req
