@@ -1,3 +1,7 @@
+const WebpackBar = require('webpackbar')
+require('dotenv').config()
+const path = process.env.NUXT_ENV || ''
+
 module.exports = {
     /*
     ** Headers of the page
@@ -23,16 +27,23 @@ module.exports = {
     css: [
         'element-ui/lib/theme-chalk/index.css',
     ],
+    env: {
+      basePath: `${path}`,
+      baseURL: (process.env.BASE_URL || 'http://localhost:3005') + `/${path}`
+    },
     /**
      * plugins
      */
     plugins: [
-        '@/plugins/element-ui'
+        '~/plugins/element-ui'
     ],
+    srcDir: 'src-nuxt/',
     /*
     ** Build configuration
     */
     build: {
+      vendor: ['babel-polyfill', 'axios', '~/plugins/element-ui.js'],
+      extractCSS: true,
       /*
       ** Run ESLint on save
       */
@@ -45,6 +56,19 @@ module.exports = {
             exclude: /(node_modules)/
           })
         }
+        const urlLoader = config.module.rules.find((rule) => rule.loader === 'url-loader')
+        config.module.rules.splice(config.module.rules.indexOf(urlLoader), 1)
+        config.module.rules.push({
+          test: /\.(png|jpe?g|gif|svg)$/,
+          loader: 'url-loader',
+          exclude: /(assets\/svg)/,
+          query: {
+            limit: 500 * 1000, // 500KO
+            name: 'img/[name].[hash:7].[ext]'
+          }
+        })
+
+        config.plugins.push(new WebpackBar())
       }
     }
   }
